+++ tests/test_parser_gitlab.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 """
 
-Copyright (C) 2020 Vanessa Sochat.
+Copyright (C) 2020-2022 Vanessa Sochat.
 
 This Source Code Form is subject to the terms of the
 Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed